chore: apply Prettier formatting on actions

+++ lib/actions/services/admin/admin.product.service.ts
@@ -3,11 +3,7 @@
 import { prisma } from "@/db/prisma";
 import { CONSTANTS, PATH, REDIS_KEY, TYPESENSE_KEY } from "@/lib/constants";
 import { redis } from "@/lib/redis";
-import {
-  cacheData,
-  deleteAllRedisKey,
-  getCachedData,
-} from "@/lib/redis/redis-handler";
+import { cacheData, deleteAllRedisKey, getCachedData } from "@/lib/redis/redis-handler";
 import { createOneIndex } from "@/lib/typesense/create-one-Index";
 import {
   ProductByTagSchemaIndexConvertor,
@@ -29,10 +25,7 @@ export type HandleGetProductType = {
   props?: Prisma.ProductFindFirstArgs;
 };
 export const handleGetProduct = async ({ id, props }: HandleGetProductType) => {
-  const product = await prisma.product.findFirst({
-    where: { id },
-    ...props,
-  });
+  const product = await prisma.product.findFirst({ where: { id }, ...props });
   if (!product) throw new Error("Product not found");
   return { data: prismaToJs(product) };
 };
@@ -48,9 +41,8 @@ export const handleGetAllAdminProduct = async ({
 }: HandleGetAllAdminProductType): Promise<{ data: AdminProductResult }> => {
   const cacheKey = `${REDIS_KEY.ADMIN_PRODUCT_LIST}_${page}`;
   const cachedList = await getCachedData<AdminProductResult>(cacheKey);
-  if (cachedList) {
-    return { data: { ...cachedList } };
-  }
+  if (cachedList) return { data: { ...cachedList } };
+
   const pagination = limit ? { skip: (page - 1) * limit } : {};
   const product = await prisma.product.findMany({
     select: {
@@ -61,9 +53,7 @@ export const handleGetAllAdminProduct = async ({
       stock: true,
       rating: true,
     },
-    orderBy: {
-      createdAt: "desc",
-    },
+    orderBy: { createdAt: "desc" },
     take: limit,
     ...pagination,
   });
@@ -84,27 +74,15 @@ export const handleCreateProduct = async (data: HandleCreateProductType) => {
   const result = await prisma.product.create({
     data: {
       ...product,
-      tags: {
-        connect: product.tags?.map((tag: TagType) => ({ id: tag.id })),
-      },
+      tags: { connect: product.tags?.map((tag: TagType) => ({ id: tag.id })) },
     },
   });
-  if (data.isFeatured) {
-    await redis.del(REDIS_KEY.BANNER);
-  }
+  if (data.isFeatured) await redis.del(REDIS_KEY.BANNER);
+
   // update-typesense
-  const updateData = {
-    ...product,
-    id: result.id,
-  };
-  await createOneIndex(
-    TYPESENSE_KEY.PRODUCT,
-    ProductSchemaIndexConvertor([updateData])
-  );
-  await createOneIndex(
-    TYPESENSE_KEY.PRODUCT_BY_TAG,
-    ProductByTagSchemaIndexConvertor([updateData])
-  );
+  const updateData = { ...product, id: result.id };
+  await createOneIndex(TYPESENSE_KEY.PRODUCT, ProductSchemaIndexConvertor([updateData]));
+  await createOneIndex(TYPESENSE_KEY.PRODUCT_BY_TAG, ProductByTagSchemaIndexConvertor([updateData]));
   await deleteAllRedisKey(REDIS_KEY.ADMIN_PRODUCT_LIST);
   return { message: "Product created successfully" };
 };
@@ -115,10 +93,7 @@ export const handleUpdateProduct = async (data: updateProductType) => {
     where: { id: data.id },
   });
   if (!originalProductData) throw new Error("Product not found");
-  if (
-    (originalProductData.isFeatured && !data.isFeatured) ||
-    (!originalProductData.isFeatured && data.isFeatured)
-  ) {
+  if ((originalProductData.isFeatured && !data.isFeatured) || (!originalProductData.isFeatured && data.isFeatured)) {
     await redis.del(REDIS_KEY.BANNER);
   }
   const product = updateProductSchema.parse(data);
@@ -126,45 +101,25 @@ export const handleUpdateProduct = async (data: updateProductType) => {
     where: { id: data.id },
     data: {
       ...product,
-      tags: {
-        set: product.tags?.map((tag) => ({ id: tag.id })),
-      },
+      tags: { set: product.tags?.map((tag) => ({ id: tag.id })) },
     },
   });
 
   // redis + typesense
-  await updateIndex(
-    TYPESENSE_KEY.PRODUCT_BY_TAG,
-    ProductByTagSchemaIndexConvertor([product])[0],
-    data.id!
-  );
-  await updateIndex(
-    TYPESENSE_KEY.PRODUCT,
-    ProductSchemaIndexConvertor([product])[0],
-    data.id!
-  );
+  await updateIndex(TYPESENSE_KEY.PRODUCT_BY_TAG, ProductByTagSchemaIndexConvertor([product])[0], data.id!);
+  await updateIndex(TYPESENSE_KEY.PRODUCT, ProductSchemaIndexConvertor([product])[0], data.id!);
   await deleteAllRedisKey(REDIS_KEY.ADMIN_PRODUCT_LIST);
   return { message: "Product updated successfully" };
 };
 
 // delete-product
 export const handleDeleteProduct = async (id: string) => {
-  const isExist = await prisma.product.findFirst({
-    where: { id },
-  });
+  const isExist = await prisma.product.findFirst({ where: { id } });
   if (!isExist) throw new Error("Product not found");
   await deleteImage(isExist.images, "product");
-  if (isExist.banner) {
-    await deleteImage([isExist.banner], "banner");
-  }
-  if (isExist.isFeatured) {
-    await redis.del(REDIS_KEY.BANNER);
-  }
-  await prisma.product.delete({
-    where: {
-      id,
-    },
-  });
+  if (isExist.banner) await deleteImage([isExist.banner], "banner");
+  if (isExist.isFeatured) await redis.del(REDIS_KEY.BANNER);
+  await prisma.product.delete({ where: { id } });
   await deleteOneItemIndex({ model: TYPESENSE_KEY.PRODUCT, id });
   await deleteOneItemIndex({ model: TYPESENSE_KEY.PRODUCT_BY_TAG, id });
   await deleteAllRedisKey(REDIS_KEY.ADMIN_PRODUCT_LIST);
